gas alert service

+++ Smart Parking Garage/Contracts/Chatbot/ChatbotMessageRequest.cs
@@ -3,6 +3,6 @@
 public class ChatbotMessageRequest
 {
     public string Message { get; set; } = string.Empty;
-    public double Latitude { get; set; }
-    public double Longitude { get; set; }
+    public float Latitude { get; set; }
+    public float Longitude { get; set; }
 }
